Refactor Playwright Tests
