+++ sample-tests/Pizzeria.Tests.Integration/HealthCheckTests.cs
@@ -1,5 +1,6 @@
 using FluentAssertions;
 using Pizzeria.Common;
+using Pizzeria.Tests.Integration.Infrastructure;
 using Xunit;
 
 namespace Pizzeria.Tests.Integration;
