+++ sample-tests/Pizzeria.Tests.Integration/Infrastructure/PageExtensions.cs
@@ -2,7 +2,7 @@ using Lewee.Playwright;
 using Microsoft.Playwright;
 using Pizzeria.Store.Web;
 
-namespace Pizzeria.Tests.Integration;
+namespace Pizzeria.Tests.Integration.Infrastructure;
 
 internal static class PageExtensions
 {
@@ -42,6 +42,6 @@ internal static class PageExtensions
 
     public static void ShouldHaveBannerHeading(this IPage page)
     {
-        page.ShouldHave(MainLayout.BannerHeading);
+        page.ShouldHave(MainLayout.Selectors.BannerHeading);
     }
 }
