+++ sample-tests/Pizzeria.Tests.Integration/Infrastructure/PizzeriaTestCollection.cs
@@ -1,0 +1,13 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using Xunit;
+
+namespace Pizzeria.Tests.Integration.Infrastructure;
+
+[SuppressMessage(
+    "Maintainability",
+    "CA1515:Consider making public types internal",
+    Justification = "xUnit requires this to be public")]
+[CollectionDefinition(PizzeriaApplicationFactory.CollectionName)]
+public sealed class PizzeriaTestCollection : ICollectionFixture<PizzeriaApplicationFactory>
+{
+}
