+++ sample/Pizzeria.Store.Web/Home.razor.cs
@@ -13,4 +13,14 @@ public partial class Home
     {
         this.Dispatcher.Dispatch(new ClearOrderErrorAction());
     }
+
+    public static class Selectors
+    {
+        public const string StartOrderButton = $"[role='button'][aria-label='{AriaLabels.StartOrder}']";
+    }
+
+    private static class AriaLabels
+    {
+        public const string StartOrder = "start order";
+    }
 }
