+++ sample/Pizzeria.Store.Web/MainLayout.razor
@@ -29,7 +29,7 @@
 
             <MudIconButton
                 Icon="@Icons.Material.Filled.Logout"
-                aria-label="sign-out"
+                aria-label="@MainLayout.AriaLabels.SignOut"
                 Color="Color.Inherit"
                 ButtonType="ButtonType.Submit" />
         </form>
