+++ sample/Pizzeria.Store.Web/MainLayout.razor.cs
@@ -4,6 +4,14 @@ namespace Pizzeria.Store.Web;
 
 public partial class MainLayout : LayoutComponentBase
 {
-    public const string BannerHeading = "[role='heading'][aria-level='1']";
-    public const string SignOutButtonSelector = "button[aria-label='sign-out']";
+    public static class Selectors
+    {
+        public const string BannerHeading = "[role='heading'][aria-level='1']";
+        public const string SignOutButton = $"button[aria-label='{AriaLabels.SignOut}']";
+    }
+
+    private static class AriaLabels
+    {
+        public const string SignOut = "sign-out";
+    }
 }
